% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{get.stars}
\alias{get.stars}
\title{Significance stars}
\usage{
get.stars(p)
}
\arguments{
\item{p}{A numeric \emph{p}-value from a hypothesis test.}
}
\value{
A string representing the significance level correponding to this
  \emph{p}-value ("\code{***}", "\code{**}", "\code{*}", "\code{.}", or "").
}
\description{
Gives the statistical significance level of a \emph{p}-value.
}
\examples{
get.stars(0.03)
}

