% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/langcog.R
\docType{package}
\name{langcog}
\alias{langcog}
\alias{langcog-package}
\title{langcog: Language and Cognition Lab Things}
\description{
The langcog package provides several different components:
\describe{ 
  \item{solarized_palette}{Colour schemes based on the solarized accent colours.} 
  \item{clm}{Linear models with coefficient constraints.}
  \item{multiboot}{Non-parametric bootstrap with multiple sample statistics.}
  \item{util}{Miscellaneous utility functions.}
}
}
