% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multiboot.R
\name{multi.boot.df}
\alias{multi.boot.df}
\title{Non-parametric bootstrap with multiple sample statistics on dataframes}
\usage{
multi.boot.df(df, col, statistics = c("ci.low", "ci.high"),
  summary.function = "mean", nboot = 1000)
}
\arguments{
\item{df}{A data frame containg the data to bootstrap.}

\item{col}{A string indicating the column to bootstrap.}

\item{statistics}{A vector of strings that are names of functions to be
computed over the set of summary values from all samples (e.g. \code{sd},
\code{ci.high}, \code{ci.low}).}

\item{summary.function}{A string that is the name of a function to be
computed over each set of samples. This function needs to take a vector and
return a single number (e.g. \code{mean}, \code{median}).}

\item{nboot}{The number of bootstrap samples to take.}
}
\description{
A wrapper for \code{multi.boot} for use with dataframes. Calls
\code{multi.boot} for a specified column and returns all statistics as their
own columns, while presevering \code{dplyr} grouping structure.
}
\examples{
## Mean and 95\% Confidence Interval for 1000 Samples from a Standard Normal
library(dplyr)
gauss.1 <- data.frame(value = rnorm(1000, mean = 0, sd = 1),
                     condition = 1)
gauss.2 <- data.frame(value = rnorm(1000, mean = 2, sd = 3),
                     condition = 2)
df <- bind_rows(gauss.1, gauss.2) \%>\% group_by(condition)
ci.low <- function(x) {quantile(x, 0.025)}
ci.high <- function(x) {quantile(x, 0.975)}
estimates <- multi.boot.df(df, "value", statistics = c("ci.low", "mean", "ci.high"))
}

