% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multiboot.R
\name{multi.boot}
\alias{multi.boot}
\title{Non-parametric bootstrap with multiple sample statistics}
\usage{
multi.boot(x, statistics = c("ci.low", "ci.high"),
  summary.function = "mean", nboot = 1000)
}
\arguments{
\item{x}{A vector of data to bootstrap over.}

\item{statistics}{A vector of strings that are names of functions to be
computed over the set of summary values from all samples (e.g. \code{sd},
\code{ci.high}, \code{ci.low}).}

\item{summary.function}{A string that is the name of a function to be
computed over each set of samples. This function needs to take a vector and
return a single number (e.g. \code{mean}, \code{median}).}

\item{nboot}{The number of bootstrap samples to take.}
}
\description{
Computes arbitrary bootstrap statistics on univariate data. By default,
computes a 95\% confidence interval of the mean of the data.
}
\examples{
## Mean and 95\% Confidence Interval for 1000 Samples from a Standard Normal
x <- rnorm(1000, mean = 0, sd = 1)
ci.low <- function(x) {quantile(x, 0.025)}
ci.high <- function(x) {quantile(x, 0.975)}
estimates <- multi.boot(x, statistics = c("ci.low", "mean", "ci.high"))
}

