% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiboot.R
\name{multi_boot_standard}
\alias{multi_boot_standard}
\title{Nonparemetric bootstrap and empirical central tendency for data frames}
\usage{
multi_boot_standard(data, col, na.rm = FALSE, empirical_function,
  statistics_functions, nboot = 1000)
}
\arguments{
\item{data}{A data frame.}

\item{col}{A string indicating the column of \code{data} to bootstrap}

\item{na.rm}{A logical indicating whether NAs should be dropped before 
bootstrapping (defaults to \code{NULL})}

\item{empirical_function}{Deprecated.}

\item{statistics_functions}{Deprecated.}

\item{nboot}{The number of bootstrap samples to take (defaults to
\code{1000}).}
}
\description{
Computes CI of bootstrapped mean of univariate data. NOTE: Bootstrapping
functions will be computed over the grouping variables currently specified
for the data frame.
}
\examples{
## Mean and 95\% confidence interval for 1000 samples from two different normal distributions
require(dplyr)
gauss1 <- data_frame(value = rnorm(1000, mean = 0, sd = 1), condition = 1)
gauss2 <- data_frame(value = rnorm(1000, mean = 2, sd = 3), condition = 2)
df <- bind_rows(gauss1, gauss2) \%>\% group_by(condition)
df \%>\% multi_boot_standard(col = "value")
}
