% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clm.R
\name{predict.clm}
\alias{predict.clm}
\title{Predict method for constrained linear model fits}
\usage{
\method{predict}{clm}(object, newdata, ...)
}
\arguments{
\item{object}{Object of class inheriting from "\code{clm}".}

\item{newdata}{A data frame in which to look for variables with which to
predict.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Predicted values based on constrained linear model object.
}

