% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clm.R
\name{predictdf.clm}
\alias{predictdf.clm}
\title{Prediction data frame for constrained linear model fits}
\usage{
predictdf.clm(model, xseq, se, level)
}
\arguments{
\item{model}{Object of class inheriting from "\code{clm}".}

\item{xseq}{See \code{\link[ggplot2]{stat_smooth}}.}

\item{se}{See \code{\link[ggplot2]{stat_smooth}}.}

\item{level}{See \code{\link[ggplot2]{stat_smooth}}.}
}
\description{
Get predicted values based on constrained linear model object into data
frame. This method is called by \code{\link[ggplot2]{stat_smooth}}.
}

