% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scale.R
\name{scale}
\alias{scale}
\title{Scale a matrix with no attributes}
\usage{
scale(x, center = TRUE, scale = TRUE)
}
\arguments{
\item{x}{the data}

\item{center}{whether you subtract the mean}

\item{scale}{whether you divide by the SD}
}
\description{
From \link[here]{http://r.789695.n4.nabble.com/How-to-remove-attributes-from-scale-in-a-matrix-td4671715.html}
}
\details{
\code{scale} is exactly the same as normal scale, except without the attributes
that make \code{\link[base]{scale}} so terrible to work with in functional
environments.
}
\examples{
scale(c(1,2,3,4,5))
}
\seealso{
\code{\link[base]{scale}}
}
\keyword{statistics}

