% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/solarized_palette.R
\name{scale_colour_solarized}
\alias{scale_color_solarized}
\alias{scale_colour_solarized}
\alias{scale_fill_solarized}
\title{solarized colour palette for ggplot2}
\usage{
scale_colour_solarized(...)

scale_color_solarized(...)

scale_fill_solarized(...)
}
\arguments{
\item{...}{Arguments passed on to discrete_scale to control name, limits,
  breaks, labels and so forth.}
}
\description{
Provides a colour scheme based on the
\href{http://ethanschoonover.com/solarized}{solarized} accent colours for use
with ggplot2.
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(x = mpg, y = hp, colour = factor(cyl))) +
  geom_point() +
  scale_colour_solarized()
}

