% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{sem}
\alias{sem}
\title{Standard error of the mean}
\usage{
sem(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector}

\item{na.rm}{Logical indicating whether missing values should be removed.}
}
\description{
Computes the standard error of the mean of the values in x. If \code{na.rm}
is \code{TRUE} then missing values are removed before computation proceeds.
}
\examples{
x <- rnorm(100)
sem(x)
}

