% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarized_palette.R
\name{solarized_palette}
\alias{solarized_palette}
\title{solarized colour palette}
\usage{
solarized_palette(num_values)
}
\arguments{
\item{num_values}{Number of values to colour. If more than 9, colours start
to repeat.}
}
\description{
Provides a colour scheme based on the
\href{http://ethanschoonover.com/solarized}{solarized} accent colours.
}
\examples{
if (requireNamespace("scales", quietly = TRUE)) {
  library(scales)
  show_col(solarized_palette(9))
}
}
