% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/solarized_palette.R
\name{theme_mikabr}
\alias{theme_mikabr}
\title{ggplot2 theme}
\usage{
theme_mikabr(base_size = 14, base_family = "Open Sans")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\description{
Has the same theme options as \code{\link{theme_bw}}, except for different
defaults for \code{base_size} and \code{base_family}, and no minor gridlines.
}
\examples{
\dontrun{
library(ggplot2)
ggplot(mtcars, aes(x = mpg, y = hp, colour = factor(cyl))) +
  geom_point() +
  theme_mikabr()
  }
}

