% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{ci_upper}
\alias{ci_upper}
\title{Confidence interval (upper)}
\usage{
ci_upper(x)
}
\arguments{
\item{x}{Numeric vector whose confidence interval is wanted.}
}
\description{
Gives the upper bound of the 95\% confidence interval of an empirical
distribution.
}
\examples{
x <- rnorm(1000)
ci_upper(x)
}

